test(glyph): keep stability regression within source limit
